Switch to user odoo to create the volumes
Else the corresponding folders will be created by root on the host.

This is a problem if the volumes have not been mapped to a folder that
belongs to the TARGET_UID user because that user won't be able to write
inside the folders (e.g. sessions, attachments, extra_addons).